+++ src/ReqRest.Client/ApiInterface.cs
@@ -1,6 +1,7 @@
 ﻿namespace ReqRest.Client
 {
     using System;
+    using System.ComponentModel;
     using ReqRest.Builders;
 
     /// <summary>
@@ -9,14 +10,14 @@
     public abstract class ApiInterface : IUrlProvider
     {
 
-        private readonly Lazy<Uri> _urlLazy;
+        private Uri? _url;
 
         /// <summary>
         ///     Gets the <see cref="ApiClient"/> which ultimately manages (or rather "contains")
         ///     this interface.
         ///     This client's configuration is supposed to be used when building requests.
         /// </summary>
-        protected ApiClient Client { get; }
+        protected internal ApiClient Client { get; }
 
         /// <summary>
         ///     Gets an <see cref="IUrlProvider"/> which is the logical parent of this
@@ -24,12 +25,12 @@
         ///     The URL which is returned by this <see cref="IUrlProvider"/> is used as this
         ///     interface's base url.
         /// </summary>
-        protected IUrlProvider BaseUrlProvider { get; }
+        protected internal IUrlProvider BaseUrlProvider { get; }
 
         /// <summary>
         ///     Gets the URL which was built for the <see cref="ApiInterface"/>.
         /// </summary>
-        protected Uri Url => _urlLazy.Value;
+        protected internal Uri Url => _url ?? (_url = this.GetUrl());
 
         /// <summary>
         ///     Initializes a new instance of the <see cref="ApiInterface"/> class whose full URL
@@ -54,7 +55,6 @@
         {
             Client = apiClient ?? throw new ArgumentNullException(nameof(apiClient));
             BaseUrlProvider = baseUrlProvider ?? apiClient;
-            _urlLazy = new Lazy<Uri>(() => this.GetUrl(), isThreadSafe: false);
         }
 
         /// <inheritdoc/>
@@ -82,7 +82,7 @@
         ///     This should usually be the incoming <paramref name="baseUrl"/> builder instance,
         ///     but can, for special cases, also be an entirely different instance.
         /// </returns>
-        protected abstract UrlBuilder BuildUrl(UrlBuilder baseUrl);
+        protected internal abstract UrlBuilder BuildUrl(UrlBuilder baseUrl);
 
         /// <summary>
         ///     Returns a new <see cref="ApiRequest"/> instance which can be used
@@ -94,17 +94,65 @@
         /// <returns>
         ///     A new <see cref="ApiRequestBase"/> instance.
         /// </returns>
-        protected virtual ApiRequest BuildRequest()
+        protected internal virtual ApiRequest BuildRequest()
         {
-            var httpClient = Client.Configuration.HttpClientProvider.Invoke();
-            return new ApiRequest(httpClient)
+            return new ApiRequest(Client.Configuration.HttpClientProvider)
                 .SetRequestUri(Url);
         }
+
+        // The following methods are overridden/shadowed, so that the EditorBrowsable Attribute
+        // can be applied.
+        // While I myself consider hiding members bad practice, I will do it in this case,
+        // because consumers of an API built via ApiInterfaces should, in my opinion,
+        // only see an IntelliSense window with the API's actual members, i.e. something like this:
+        //  ________                        ______________
+        // | Get    |                      | Get          |
+        // | Post   |   instead of this:   | GetHashCode  |
+        // | Items  |                      | GetType      |
+        // |________|                      | Post         |
+        //                                 | ...          |
+        //                                 |______________|
+        //
+        // It appears that this doesn't work for GetType(), but we will still keep it here,
+        // for consistency. Furthermore, it may work as intended in SOME editors. Who knows.
+        // Let's try to get the most out of it.
+
+        /// <summary>
+        ///     Gets the <see cref="Type"/> of the current instance.
+        ///     This method calls the <see cref="object.GetType"/> method and returns its result.
+        /// </summary>
+        /// <returns>The exact runtime type of the current instance.</returns>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public new Type GetType() =>
+            base.GetType();
+
+        /// <summary>
+        ///     Determines whether the specified object is equal to the current object.
+        ///     This method calls the <see cref="object.Equals(object)"/> method and returns its result.
+        /// </summary>
+        /// <param name="obj">The object to compare with the current object.</param>
+        /// <returns>
+        ///     <see langword="true"/> if the specified object is equal to the current object;
+        ///     otherwise, <see langword="false"/>.
+        /// </returns>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override bool Equals(object obj) =>
+            base.Equals(obj);
+
+        /// <summary>
+        ///     Serves as the default hash function.
+        ///     This method calls the <see cref="object.GetHashCode"/> method and returns its result.
+        /// </summary>
+        /// <returns>A hash code for the current object.</returns>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override int GetHashCode() =>
+            base.GetHashCode();
 
         /// <summary>
         ///     Returns a string representing the interface's URL.
         /// </summary>
         /// <returns>A string representing the interface's URL.</returns>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public override string ToString() =>
             Url.ToString();
 
@@ -128,7 +176,7 @@
         ///     this interface.
         ///     This client's configuration is supposed to be used when building requests.
         /// </summary>
-        protected new TClient Client => (TClient)base.Client;
+        protected internal new TClient Client => (TClient)base.Client;
 
         /// <summary>
         ///     Initializes a new instance of the <see cref="ApiInterface"/> class whose full URL