+++ src/ReqRest.Client/ApiResponseBase.cs
@@ -43,12 +43,11 @@
         /// <summary>
         ///     Gets a <see cref="ResponseTypeInfo"/> from the <see cref="PossibleResponseTypes"/>
         ///     set which is the most appropriate one for the response's status code.
-        ///     Deriving classes can perform deserialization logic based on this information.
         ///     
-        ///     By default, this is <see langword="null"/> if the <see cref="PossibleResponseTypes"/>
+        ///     This is <see langword="null"/> if the <see cref="PossibleResponseTypes"/>
         ///     property doesn't contain any information which matches the response's status code.
         /// </summary>
-        protected virtual ResponseTypeInfo? CurrentResponseTypeInfo { get; }
+        protected ResponseTypeInfo? CurrentResponseTypeInfo { get; }
 
         /// <summary>
         ///     Initializes a new <see cref="ApiResponseBase"/> instance with the specified values.
@@ -68,7 +67,7 @@
             IEnumerable<ResponseTypeInfo>? possibleResponseTypes)
             : base(httpResponseMessage)
         {
-            PossibleResponseTypes = possibleResponseTypes ?? Enumerable.Empty<ResponseTypeInfo>();
+            PossibleResponseTypes = possibleResponseTypes?.ToList() ?? Enumerable.Empty<ResponseTypeInfo>();
             CurrentResponseTypeInfo = FindMostAppropriateResponseTypeInfo();
         }
 
@@ -99,23 +98,8 @@
         ///     <see langword="false"/> if not.
         /// </returns>
         private protected virtual bool CanDeserializeResource<T>() =>
-            CanDeserializeResource(typeof(T));
-
-        /// <summary>
-        ///     Returns a value indicating whether a resource of the specified <paramref name="resourceType"/>
-        ///     can be deserialized from this response's content.
-        ///     The value is determined based on <see cref="CurrentResponseTypeInfo"/>.
-        /// </summary>
-        /// <param name="resourceType">
-        ///     The type of the resource to be deserialized.
-        /// </param>
-        /// <returns>
-        ///     <see langword="true"/> if a resource of the specified type can be deserialized;
-        ///     <see langword="false"/> if not.
-        /// </returns>
-        private protected virtual bool CanDeserializeResource(Type resourceType) =>
             CurrentResponseTypeInfo != null &&
-            resourceType.IsAssignableFrom(CurrentResponseTypeInfo.ResponseType);
+            typeof(T).IsAssignableFrom(CurrentResponseTypeInfo.ResponseType);
 
         /// <summary>
         ///     Attempts to deserialize the response to the specified type <typeparamref name="T"/>
@@ -145,7 +129,7 @@
             {
                 return await deserializer.DeserializeAsync<T>(HttpResponseMessage.Content).ConfigureAwait(false);
             }
-            catch (Exception ex) when(!(ex is HttpContentSerializationException))
+            catch (Exception ex) when (!(ex is HttpContentSerializationException))
             {
                 // Ideally, the deserializer throws this exception by himself, but we cannot count on that.
                 throw new HttpContentSerializationException(null, ex);